fixed encryption

+++ src/Pair.java
@@ -1,18 +1,18 @@
 
 public class Pair {
-	private int val1;
-	private int val2;
-	public Pair(int x, int y) {
+	private long val1;
+	private long val2;
+	public Pair(long x, long y) {
 		val1 = x;
 		val2 = y;
 	}
-	public int getVal1() {
+	public long getVal1() {
 		return val1;
 	}
 	public void setVal1(int val1) {
 		this.val1 = val1;
 	}
-	public int getVal2() {
+	public long getVal2() {
 		return val2;
 	}
 	public void setVal2(int val2) {
